elf: Move Elf64Shdr to separate module
In the effort of splitting up the ELF parser code, move the Elf64Shdr
implementation to its own module.
